requirements.txt
# Core dependencies
httpx>=0.25.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
opencv-python>=4.8.0
replicate>=0.15.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0

# Optional dependencies for better performance
lxml>=4.9.0

# Background removal dependencies
scipy>=1.9.0